pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "larasanic"
version = "0.1.0"
description = "A Laravel-inspired async web framework built on Sanic"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Larasanic Team"}
]
keywords = ["web", "framework", "async", "sanic", "laravel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: AsyncIO",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    # Core Web Framework
    "sanic>=23.0.0",
    "sanic-ext>=23.0.0",

    # Database ORM & Migrations
    "tortoise-orm>=0.19.0",
    "aerich>=0.7.0",
    "asyncpg>=0.27.0",
    "aiomysql>=0.1.0",

    # HTTP Client
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "certifi>=2022.0.0",

    # Async Redis (for cache and websocket support)
    "redis>=4.5.0",

    # Security & Crypto
    "bcrypt>=4.0.0",
    "PyJWT>=2.6.0",
    "cryptography>=40.0.0",
    "itsdangerous>=2.1.0",

    # Environment & Configuration
    "python-dotenv>=0.21.0",

    # Utilities
    "tomlkit>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/larasanic"
Documentation = "https://larasanic.readthedocs.io"
Repository = "https://github.com/yourusername/larasanic"
Changelog = "https://github.com/yourusername/larasanic/blob/main/CHANGELOG.md"

[project.scripts]
larasanic = "larasanic.console.artisan:main"

[tool.setuptools]
packages = ["larasanic"]

[tool.setuptools.package-data]
larasanic = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
